[feat] : Validation complete for form

+++ src/actions/actions.ts
@@ -1,15 +1,21 @@
 "use server";
 import prisma from "@/lib/db";
-import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
-import { Pet } from "@prisma/client";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(pet);
+  if(!validatedPet.success) {
+    return {
+      message: "Invalid pet data",
+    }
+  }
   try {
     await prisma.pet.create({
-      data: pet
+      data: validatedPet.data
     })
   } catch (error) {
     console.error(error);
@@ -20,14 +26,22 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath("/app", 'layout');
 }
 
-export async function editPet(petId: Pet['id'], newPet: PetEssentials) {
+export async function editPet(petId: unknown, newPet: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPet);
+  if(!validatedPetId.success || !validatedPet.success) {
+    return {
+      message: "Invalid pet data",
+    }
+  }
   try {
     await prisma.pet.update({
       where: {
-        id: petId
+        id: validatedPetId.data
       },
-      data: newPet
+      data: validatedPet.data
     })
   } catch (error) {
     console.error(error);
@@ -38,12 +52,18 @@ export async function editPet(petId: Pet['id'], newPet: PetEssentials) {
   revalidatePath("/app", 'layout');
 } 
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if(!validatedPetId.success) {
+    return {
+      message: "Invalid pet id",
+    }
+  }
   try {
     await prisma.pet.delete({
       where: {
-        id: petId
+        id: validatedPetId.data
       }
     })
   } catch (error) {
